[Refactor] useWindowWidthSize 리팩터링 > useGetDiviceType > 디바이스 타입만 반환하도록 변경
