+++ src/hooks/useGetDeviceType.js
@@ -1,12 +1,7 @@
 import { useState, useEffect } from "react";
 
-const useWindowWidhtSize = () => {
+const useGetDeviceType = () => {
   const [windowWidhth, setWindowWidthSize] = useState(window.innerWidth);
-
-  const [bestProductCount, setBestProductCount] = useState(4);
-  const [sellingProductCount, setSellingProductCount] = useState(10);
-  const [sellingProductCountPerRow, setSellingProductCountPerRow] = useState(5);
-
   const [device, setDevice] = useState("PC");
 
   useEffect(() => {
@@ -24,19 +19,10 @@ const useWindowWidhtSize = () => {
 
   useEffect(() => {
     if (windowWidhth > 800) {
-      setSellingProductCount(10);
-      setSellingProductCountPerRow(5)
-      setBestProductCount(4);
       setDevice("PC")
-    } else if (windowWidhth <= 800 && windowWidhth > 400) {
-      setSellingProductCount(6);
-      setSellingProductCountPerRow(3)
-      setBestProductCount(2);
+    } else if (windowWidhth <= 744 && windowWidhth > 376) {
       setDevice("Tablet")
-    } else if (windowWidhth <= 400) {
-      setSellingProductCount(4);
-      setSellingProductCountPerRow(2)
-      setBestProductCount(1);
+    } else if (windowWidhth <= 375) {
       setDevice("Mobile")
     }
     console.log(device)
@@ -47,4 +33,4 @@ const useWindowWidhtSize = () => {
 
 };
 
-export default useWindowWidhtSize;
+export default useGetDeviceType;
