+++ src/pages/Homepage.js
@@ -9,7 +9,7 @@ import SellingProductRender from "components/SellingProductRender";
 
 // 커스텀 훅
 import useProductData from "../hooks/useProductData.js";
-import useWindowWidhtSize from "../hooks/useWindowWidhtSize.js";
+import useWindowWidhtSize from "../hooks/useGetDeviceType.js/index.js";
 
 function Hompage() {
   const [ProductSortOption, setProductSortOption] = useState("recent");
