+++ src/pages/_Homepage copy.js
@@ -3,12 +3,12 @@ import { useState } from "react";
 import styles from "./Homepage.module.css";
 
 //렌더링 컴포넌트
-import PageNav from "../components/PageNav.js"
+import PageNav from "../components/PageNav.js";
 import SellingProductHeader from "../components/SellingProductHeader.js";
 
 // 커스텀 훅
 import useProductData from "../hooks/useProductData.js";
-import useWindowWidhtSize from "../hooks/useWindowWidhtSize.js";
+import useWindowWidhtSize from "../hooks/useGetDeviceType.js/index.js";
 
 function Hompage() {
   const [ProductSortOption, setProductSortOption] = useState("recent");
@@ -44,48 +44,30 @@ function Hompage() {
 
   return (
     <div className={styles.homepage}>
-      <PageNav />
-      <div>
-        {/* <SellingProductHeader text={"판매 중인 상품"} handleSortOption={handleSortOption} devicetype={device}/> */}
-      </div>
-      {/* <HomepageRenderFooter
-        nowPage={nowPage}
-        handlePageChange={handlePageChange}
-        totalPageSize={totalPageSize}
-      /> */}
+      <section className="bestProductSection">
+        <header className="ProductSectionHeader">
+          <ProductHeaderText headerText={"베스트 상품"} />
+        </header>
+        <ProductRenderGrid
+          productData={bestProductData}
+          noProduct={bestNoProduct}
+        />
+      </section>
+      <section className="sellingProductSection">
+        <header className="ProductSectionHeader">
+          <ProductHeaderText headerText={"판매 중인 상품"} />
+          <ProductHeaderSearchBar
+            inputText={searchKeyword}
+            handleInput={handleSeachKeyword}
+            device={Device}
+          />
+          <ProductHeaderRegistBtn />
+          <ProductHeaderSortBtn handleSortOption={handleSetProductSortOption} />
+        </header>
+
+      </section>
     </div>
   );
 }
 
 export default Hompage;
-
-
-// <section className="bestProductSection">
-// <header className="ProductSectionHeader">
-//   <ProductHeaderText headerText={"베스트 상품"} />
-// </header>
-// <ProductRenderGrid
-//   productData={bestProductData}
-//   noProduct={bestNoProduct}
-// />
-// </section>
-// <section className="sellingProductSection">
-// <header className="ProductSectionHeader">
-//   <ProductHeaderText headerText={"판매 중인 상품"} />
-//   <ProductHeaderSearchBar
-//     inputText={searchKeyword}
-//     handleInput={handleSeachKeyword}
-//     device={Device}
-//   />
-//   <ProductHeaderRegistBtn />
-//   <ProductHeaderSortBtn
-//     handleSortOption={handleSetProductSortOption}
-//   />
-// </header>
-// <ProductRenderGrid
-//   productData={sellingProductData}
-//   productRowCount={2}
-//   productCountPerRow={sellingProductCountPerRow}
-//   noProduct={sellingNoProduct}
-// />
-// </section>
